+++ frontend/srcs/src/app/modules/challenge/components/base/challenge.component.ts
@@ -1,19 +1,29 @@
-import { Component, TemplateRef, inject } from '@angular/core';
+import { Component, OnInit, TemplateRef, inject } from '@angular/core';
 import { ModalDismissReasons, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ChallengeInfoComponent } from '../challenge-info/challenge-info.component';
+import { ChallengeService } from 'src/app/services/challenge.service';
+import { AuthService } from 'src/app/services';
 
 @Component({
   selector: 'app-challenge',
   templateUrl: './challenge.component.html',
   styleUrl: './challenge.component.scss'
 })
-export class ChallengeComponent{
+export class ChallengeComponent implements OnInit{
   private modalService = inject(NgbModal);
+  private challengeService = inject(ChallengeService);
+  private authService = inject(AuthService);
   closeResult = '';
 
+  ngOnInit(): void {
+    this.challengeService.hereComesANewChallenger(this.authService.getMyUserId()).subscribe(newChallengerUserId => {
+      this.openModal(newChallengerUserId);
+    });
+  }
 
-  openModal() {
+  openModal(newChallengerUserId: string) {
     const modalRef = this.modalService.open(ChallengeInfoComponent);
+    modalRef.componentInstance.newChallengerUserId = newChallengerUserId;
     // Optional: Pass data to the modal component if needed
     // modalRef.componentInstance.someData = someValue;