+++ frontend/srcs/src/app/modules/challenge/components/challenge-info/challenge-info.component.ts
@@ -1,31 +1,35 @@
-import { Component, Input, inject } from '@angular/core';
+import { Component, Input, OnInit, inject } from '@angular/core';
+import { UsersCache } from 'src/app/cache';
 import { User } from 'src/app/models';
 import { ChallengeService } from 'src/app/services/challenge.service';
-// import { CommonModule } from '@angular/common';
 
 @Component({
   selector: 'app-challenge-info',
-  // standalone: true,
-  // imports: [CommonModule],
   templateUrl: './challenge-info.component.html',
   styleUrl: './challenge-info.component.scss'
 })
-export class ChallengeInfoComponent  {
-  @Input() userId!: string;
+export class ChallengeInfoComponent implements OnInit {
+  @Input() newChallengerUserId!: string;
   user: User = {
     userId: '42',
     username: 'ramdon',
     avatar: 'https://api.dicebear.com/7.x/bottts/svg?seed=ramdon'
   }
 
+  private cachedUsers = inject(UsersCache);
+
+  ngOnInit(): void {
+    this.user = this.cachedUsers.getUser(this.newChallengerUserId);
+  }
+
   private challengeService = inject(ChallengeService);
 
   acceptChallenge(){
-    this.challengeService.acceptChallenge(this.userId);
+    this.challengeService.acceptChallenge(this.newChallengerUserId);
   }
   
   rejectChallenge(){
-    this.challengeService.rejectChallenge(this.userId);
+    this.challengeService.rejectChallenge(this.newChallengerUserId);
   }
 
 }